refactor: minor shifting around of functionality
