+++ src/main.rs
@@ -1,5 +1,9 @@
-use anyhow::{bail, Result};
-use std::path::Path;
+use anyhow::{bail, Context, Result};
+use std::{
+    env, fs,
+    hash::{DefaultHasher, Hasher},
+    path::{Path, PathBuf},
+};
 use tui::Tui;
 
 use clap::Parser;
@@ -14,16 +18,21 @@ mod data_collection;
 mod harness;
 mod runners;
 mod session;
-use crate::{afl_cmd_gen::AFLCmdGenerator, harness::Harness, runners::runner::Runner};
+use crate::{
+    afl_cmd::{Printable, ToStringVec},
+    afl_cmd_gen::AFLCmdGenerator,
+    cli::{Config, RunArgs},
+    harness::Harness,
+    runners::runner::Runner,
+};
 use crate::{cli::AFL_CORPUS, runners::screen::Screen};
 use crate::{runners::tmux::Tmux, session::CampaignData};
 mod log_buffer;
 mod seed;
 mod system_utils;
 mod tui;
-mod utils;
 
-use utils::{generate_session_name, load_config};
+pub static DEFAULT_AFL_CONFIG: &str = "aflr_cfg.toml";
 
 fn main() -> Result<()> {
     let cli_args = Cli::parse();
@@ -34,6 +43,43 @@ fn main() -> Result<()> {
         Commands::Kill(kill_args) => handle_kill_command(&kill_args)?,
     }
     Ok(())
+}
+
+/// Loads the configuration from the specified `config_path` or the default configuration file.
+///
+/// If `config_path` is `None`, the function looks for a default configuration file named `aflr_cfg.toml`
+/// in the current working directory. If the default configuration file is not found, an empty `Config`
+/// instance is returned.
+///
+/// # Errors
+///
+/// Returns an error if the configuration file cannot be read or parsed.
+pub fn load_config(config_path: Option<&PathBuf>) -> Result<Config> {
+    if let Some(path) = config_path {
+        let config_content = fs::read_to_string(path)
+            .with_context(|| format!("Failed to read config file: {}", path.display()))?;
+        toml::from_str(&config_content)
+            .with_context(|| format!("Failed to parse config file: {}", path.display()))
+    } else {
+        let cwd = env::current_dir().context("Failed to get current directory")?;
+        let default_config_path = cwd.join(DEFAULT_AFL_CONFIG);
+        if default_config_path.exists() {
+            let config_content = fs::read_to_string(&default_config_path).with_context(|| {
+                format!(
+                    "Failed to read default config file: {}",
+                    default_config_path.display()
+                )
+            })?;
+            toml::from_str(&config_content).with_context(|| {
+                format!(
+                    "Failed to parse default config file: {}",
+                    default_config_path.display()
+                )
+            })
+        } else {
+            bail!("No config file provided and no default configuration file in CWD found")
+        }
+    }
 }
 
 fn load_merged_gen_args_and_flags(
@@ -102,9 +148,9 @@ fn create_afl_runner(
 
 fn handle_gen_command(gen_args: &cli::GenArgs) -> Result<()> {
     let (merged_args, raw_afl_flags) = load_merged_gen_args_and_flags(gen_args)?;
-    let mut afl_runner = create_afl_runner(&merged_args, raw_afl_flags, false)?;
-    let cmds = afl_runner.generate()?;
-    utils::print_generated_commands(&cmds);
+    let afl_runner = create_afl_runner(&merged_args, raw_afl_flags, false)?;
+    let cmds = afl_runner.run()?;
+    cmds.print();
     Ok(())
 }
 
@@ -113,11 +159,11 @@ fn handle_run_command(run_args: &cli::RunArgs) -> Result<()> {
     if merged_args.tui && merged_args.detached {
         bail!("TUI and detached mode cannot be used together");
     }
-    let mut afl_runner =
+    let afl_runner =
         create_afl_runner(&merged_args.gen_args, raw_afl_flags, merged_args.is_ramdisk)?;
-    let afl_cmds = afl_runner.generate()?;
+    let afl_cmds = afl_runner.run()?;
     if merged_args.dry_run {
-        utils::print_generated_commands(&afl_cmds);
+        afl_cmds.print();
         return Ok(());
     }
     let target_args = merged_args
@@ -131,8 +177,10 @@ fn handle_run_command(run_args: &cli::RunArgs) -> Result<()> {
     let pid_fn = format!("/tmp/.{}_{}.pids", &sname, std::process::id());
     let pid_fn_path = Path::new(&pid_fn);
     let srunner: Box<dyn Runner> = match &merged_args.session_runner {
-        SessionRunner::Screen => Box::new(Screen::new(&sname, &afl_cmds, pid_fn_path)),
-        SessionRunner::Tmux => Box::new(Tmux::new(&sname, &afl_cmds, pid_fn_path)),
+        SessionRunner::Screen => {
+            Box::new(Screen::new(&sname, &afl_cmds.to_string_vec(), pid_fn_path))
+        }
+        SessionRunner::Tmux => Box::new(Tmux::new(&sname, &afl_cmds.to_string_vec(), pid_fn_path)),
     };
 
     if merged_args.tui {
@@ -171,6 +219,39 @@ fn validate_tui_output_dir(output_dir: &Path) -> Result<()> {
         }
     }
     Ok(())
+}
+
+/// Generates a unique session name based on the provided `RunArgs` and `target_args`.
+///
+/// If the `session_name` is not specified in `RunArgs`, the function generates a unique name
+/// by combining the target binary name, input directory name, and a hash of the `target_args`.
+pub fn generate_session_name(args: &RunArgs, target_args: &str) -> String {
+    args.session_name.as_ref().map_or_else(
+        || {
+            let target = args
+                .gen_args
+                .target
+                .as_ref()
+                .expect("Target binary is required")
+                .file_name()
+                .unwrap_or_default()
+                .to_string_lossy();
+            let to_hash = format!(
+                "{}_{}_{}",
+                target,
+                args.gen_args.input_dir.as_ref().map_or_else(
+                    || AFL_CORPUS.into(),
+                    |dir| dir.file_name().unwrap_or_default().to_string_lossy()
+                ),
+                target_args,
+            );
+            let mut hasher = DefaultHasher::new();
+            hasher.write(to_hash.as_bytes());
+            let hash = hasher.finish() % 1_000_000;
+            format!("{target}_{hash}")
+        },
+        std::clone::Clone::clone,
+    )
 }
 
 fn handle_kill_command(kill_args: &cli::KillArgs) -> Result<()> {
